Removes watching, improves filtering.
Watching doesn't work reliably/consistently across all platforms in NodeJS and I suspect no one was actually using it.
On top of that, it was using a lot of node sync calls which isn't great and it was preventing improvements to the filter function.

The filter function now is asynchronous, so you can have the time to read the file/metadata before making a decision on whether to skip or not.
It also includes information on whether the "file" is actually a directory so you can skip/not-skip whole directories easily.

Updated TypeScript to latest, added declaration maps.